CompAct changes
Changed the minimum value of the 'metric' parameter in rw from 1 to 0.1.  In the 'from_aux_norm' 'fbase' and 'lrate' functions, made it so that the minimum value of 'atn_gain' on each trial is 0.01 (as in the R code for simulating CompAct).  Removed the upper and lower limits on the 'from_aux_norm' lrates.