rationalized simulation output
Changed the way that exemplar information is stored in trials objects to be the same as it is stored in exemplar model output.  Instead of creating a new dataset from scratch when simulating a model, made it so that the trials dataset is the basis of the simulation dataset and then additional simulation data is added.