changed how the name attribute works
Changed how the name attribute works for stage and schedule objects.  Instead of specifying the name when creating a stage or schedule, the default name attribute is None.  When a schedule is created, its stages get names based on dict keys.  Similarly when an experiment is created its schedules get names based on dict keys.  This greatly simplifies how experiment objects etc. work.